+++ data/website_metadata/conspiracy.json
@@ -272,6 +272,10 @@
         "url": "https://www.Bugout.new"
     },
     {
+        "name": "CaliforniaCollapse.Ne",
+        "url": "https://www.CaliforniaCollapse.New"
+    },
+    {
         "name": "CampusInsanity.c",
         "url": "https://www.CampusInsanity.co"
     },
