+++ data/website_metadata/science.json
@@ -292,6 +292,10 @@
         "url": "https://www.mdpi.com/journal/biomimetics"
     },
     {
+        "name": "Biomolecules Journal",
+        "url": "https://www.mdpi.com/journal/biomolecules"
+    },
+    {
         "name": "BioSpace",
         "url": "https://www.biospace.com"
     },
@@ -728,6 +732,10 @@
         "url": "https://www.frontiersin.org/journals/developmental-psychology"
     },
     {
+        "name": "Frontiers in Digital Health",
+        "url": "https://www.frontiersin.org/journals/digital-health"
+    },
+    {
         "name": "Frontiers in Psychology",
         "url": "https://www.journal.frontiersin.org"
     },
@@ -910,6 +918,10 @@
     {
         "name": "Live Science",
         "url": "https://www.livescience.com"
+    },
+    {
+        "name": "Malaria Journal",
+        "url": "https://www.malariajournal.biomedcentral.com"
     },
     {
         "name": "Matter Journal",
