+++ src/app/static/chartjs-plugin-datalabels@2.0.js
