+++ src/app/static/github-logo-dark.png
