+++ src/app/static/github-logo-light.png
