+++ src/app/static/reddit-logo.png
