--- test.py
@@ -1,5 +1,0 @@
-
-from text_processor import TextProcessor
-
-a = TextProcessor()
-print(a.__dict__)
